fix dynamicvoicechannel remove other channel

+++ datapackage/botfunction/dynamicVoiceChannel.js
@@ -40,23 +40,26 @@ client.on('voiceStateUpdate', async (oldState, newState) => {
             });
 
             voiceStateCollector.on('collect', async (state) => {
-                // 確認當前狀態是否是離開且此時頻道無其他成員
-                if (state.channel && state.channel.members.size === 0) {
-                    await channel.delete(); // 刪除頻道
+                // 确认当前状态是否是离开且此时频道无其他成员，并且是由该用户创建的频道
+                if (state.channel && state.channel.members.size === 0 && state.channel.name === `🛵${member.user.username}'s Channel`) {
+                    await channel.delete(); // 删除频道
                     voiceStateCollector.stop(); // 停止收集器
                 }
-            });            
+            });
+            
         } catch (error) {
             console.error('Failed to create the channel:', error);
         }
     }
 
-    // 檢查是否有成員離開了之前創建的語音頻道
+    // 检查是否有成员离开了之前创建的语音频道
     if (oldState.channel && oldState.channel.members.size === 0 && oldState.channelId !== triggerChannelId) {
-        await oldState.channel.delete(); // 刪除空的語音頻道
+        // 确保只删除由该用户创建的专属语音频道
+        if (oldState.channel.name === `🛵${oldState.member.user.username}'s Channel`) {
+            await oldState.channel.delete(); // 删除空的语音频道
+        }
     }
-});
-}
+})}
 
 export {
     dynamicvoicechannel,
